agent/requirements.txt
# LiveKit Agents dependencies
livekit-agents[silero,turn-detector]~=1.3
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
aiohttp~=3.9


